+++ GeneralsMD/Code/Libraries/Source/WWVegas/WWLib/thread.h
@@ -28,6 +28,7 @@
 
 #include "always.h"
 #include "vector.h"
+#include <stdint.h>
 
 struct _EXCEPTION_POINTERS;
 
@@ -99,7 +100,7 @@ protected:
 
 private:
 	static void __cdecl Internal_Thread_Function(void*);
-	volatile unsigned long handle;
+	volatile uintptr_t handle;
 	int thread_priority;
 };
